Create view and viewmodel layers for basic main view implementation
The MainActivity view is created, along with a RecyclerAdapter for the main list.

Features implemented in this commit (sans real model) include:
- Main view entries list (as a RecyclerView)
- Sorting
- Searching
- Navigation
- Menus of entries/sorting
- Sticky headers/dividers

Features not yet implemented (at the view layer) include:
- Password editing/creation dialogs/screens
- Hierarchical navigation (e.g. clicking folders)
- Configure screen
- Closing database

The model remains a 'dummy' model, although some basic beginnings of the model
are implemented as necessary for a working build.

This commit marks a functional prototype demonstrating the view.