Implement storage file chooser in creation screen and creation screen validation
The creation screen now has functional options to choose files for cloud or device storage. These
open dialogs to create a file in Drive, or to create a new file on the device. This latter change
required a new layout (component_file_create_dialog.xml) for the dialog.

Basic validation is also implemented in the creation screen, including (hard is enforced, soft is
presented as a warning to the user):
- Master password emptiness [hard]
- Master password length [hard]
- Master password complexity [soft]
- Master password blacklist-state [soft]
- Master password re-entry [hard]
- File selection (must select file for the chosen option) [hard]
- Chosen device file name (must be valid) [hard]

To store required variables for a single database session, a new package is added relating to the
session. This package is currently incomplete, as the remainder of the 'model' (i.e. the crypto
facilities) are not yet implemented in the project.

The purpose of this implementation at this stage is to allow for the testing and demonstration of
the storage functionality in the application. It is still not yet functional in the broader context.