Create EncryptedDatabaseSerializer & implementations
'EncryptedDatabaseSerializer' is added to allow for the serialization of a PasswordDatabase to a
binary format, the encryption of this, and the bundling of necessary decryption parameters.

The main implementation is VersionedEncryptionDatabase, which reads bytes of the file to determine
the file format version, and chooses an implementation based on this.