Create PasswordFlagsService and refactor use of SecurityServices
A new service, PasswordFlagsService, is created to allow for calculation of password flags. This is backed by a blacklist of passwords for a particular flag, which exists as an SQLite database, bundled with the application in the res/raw directory.

Scripts are added to generate the SQLite database from a wordlist. It is indexed using a dual index of first character of the password and length of the password.

SecurityServices is no longer passed to the creation VM on instantiation, due to its preferred use as a singleton. This will be the canonical use in code from now on.